Add R test with exact result, including paying attention to rounding
Adapted Mostafa's method, with two changes:

1. Instead of precomputing the binomial co-efficients, compute them
recursively as needed, using functools.lru_cache for caching the results
of function calls (this is equivalent to dynamic programming, but allows
for much simpler code)

2. Instead of computing with (n over k) * 2**(-n), which can become 0
for large n (e.g. for n = 1075 and k = 0), multiply all these numbers
with a larger constant (e.g. 2**10000) and work with integer division.
This works because Python supports integer arithmetic with unlimited
precision